requirements.txt
click
pyspellchecker
ujson
nltk
opencv-python
pytesseract